# docs/openapi/b1.yaml
openapi: 3.1.0
info:
  title: LapanganGo Phase B2 API
  version: 2.0.0
  description: Kontrak kumulatif B1 dan B2 untuk runtime lokal sandbox LapanganGo.
...
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
paths:
...
  /media:
    get:
      tags: [Venues]
      operationId: getPublicMedia
      parameters:
        - name: key
          in: query
          required: true
          schema:
            type: string
            minLength: 10
            maxLength: 255
      responses:
        "302":
          description: Redirect to a short-lived download URL for a public media asset.
          headers:
            Location:
              required: true
              schema: { type: string, format: uri }
            Cache-Control:
              schema: { type: string }
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /business/workspaces:
    get:
      tags: [Business]
      operationId: listWorkspaces
      responses:
        "200":
          description: Active business workspaces for the current identity.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Workspace" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /business/tenants:
    post:
      tags: [Business]
      operationId: createTenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 3, maxLength: 80 }
      responses:
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "201":
          description: Draft tenant and Primary Owner membership created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
  /business/venues/{venueId}/media/signed-upload:
    post:
      tags: [Business]
      operationId: createVenueSignedUpload
      parameters:
        - $ref: "#/components/parameters/VenueId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSignedUploadInput" }
      responses:
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "201":
          description: Short-lived object storage upload URL scoped to an Owner venue.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SignedUpload" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /business/tenants/{tenantId}/members:
    get:
      tags: [Business]
      operationId: listWorkspaceMembers
      parameters:
        - $ref: "#/components/parameters/TenantId"
      responses:
        "200":
          description: Tenant membership and Staff assignments.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/WorkspaceMember" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Business]
      operationId: addWorkspaceMember
      parameters:
        - $ref: "#/components/parameters/TenantId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, role]
              properties:
                email: { type: string, format: email }
                role: { type: string, enum: [OWNER, STAFF] }
      responses:
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "201":
          description: Existing registered user added to workspace.
          content:
            application/json:
              schema:
                type: object
                required: [membershipId]
                properties:
                  membershipId: { $ref: "#/components/schemas/PublicId" }
        "409": { $ref: "#/components/responses/Conflict" }
  /business/tenants/{tenantId}/staff/{membershipId}/assignments:
    put:
      tags: [Business]
      operationId: updateStaffAssignments
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - name: membershipId
          in: path
          required: true
          schema: { $ref: "#/components/schemas/PublicId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [venueIds, reason]
              properties:
                venueIds:
                  type: array
                  items: { $ref: "#/components/schemas/PublicId" }
                reason: { type: string, minLength: 3, maxLength: 500 }
      responses:
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "204": { description: Assignments replaced. }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /business/venues/{venueId}/media:
    post:
      tags: [Business]
      operationId: completeVenueMedia
      parameters:
        - $ref: "#/components/parameters/VenueId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CompleteVenueMediaInput" }
      responses:
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "201":
          description: Uploaded asset linked to venue.
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id: { $ref: "#/components/schemas/PublicId" }
...
  /business/tenants/{tenantId}/roles:
    parameters:
      - {
          name: tenantId,
          in: path,
          required: true,
          schema: { $ref: "#/components/schemas/PublicId" },
        }
    get:
      tags: [Business]
      operationId: listTenantRoles
      responses:
        "200":
          description: Tenant-scoped roles.
          content:
            {
              application/json:
                {
                  schema:
                    {
                      type: object,
                      properties:
                        {
                          items:
                            {
                              type: array,
                              items: { $ref: "#/components/schemas/TenantRole" },
                            },
                        },
                    },
                },
            }
    post:
      tags: [Business]
      operationId: copyTenantRoleTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [templateId, name, reason]
              properties:
                templateId: { $ref: "#/components/schemas/PublicId" }
                name: { type: string }
                reason: { type: string, minLength: 3, maxLength: 500 }
      responses:
        "201":
          {
            description: Role copied,
            content:
              {
                application/json: { schema: { $ref: "#/components/schemas/B2Object" } },
              },
          }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
  /business/tenants/{tenantId}/roles/{roleId}/permissions:
    put:
      tags: [Business]
      operationId: updateTenantRolePermissions
      parameters:
        - {
            name: tenantId,
            in: path,
            required: true,
            schema: { $ref: "#/components/schemas/PublicId" },
          }
        - {
            name: roleId,
            in: path,
            required: true,
            schema: { $ref: "#/components/schemas/PublicId" },
          }
      requestBody:
        required: true
        content:
          {
            application/json:
              {
                schema:
                  {
                    type: object,
                    required: [permissions, reason],
                    properties:
                      {
                        permissions: { type: array, items: { type: string } },
                        reason: { type: string },
                      },
                  },
              },
          }
      responses:
        "204": { description: Permissions updated. }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
  /business/tenants/{tenantId}/staff/{membershipId}/role:
    put:
      tags: [Business]
      operationId: assignTenantRole
      parameters:
        - {
            name: tenantId,
            in: path,
            required: true,
            schema: { $ref: "#/components/schemas/PublicId" },
          }
        - {
            name: membershipId,
            in: path,
            required: true,
            schema: { $ref: "#/components/schemas/PublicId" },
          }
      requestBody:
        required: true
        content:
          {
            application/json:
              {
                schema:
                  {
                    type: object,
                    required: [roleId, reason],
                    properties:
                      {
                        roleId: { $ref: "#/components/schemas/PublicId" },
                        reason: { type: string, minLength: 3, maxLength: 500 },
                      },
                  },
              },
          }
      responses:
        "204": { description: Role assigned. }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
...
components:
  parameters:
...
    TenantId:
      name: tenantId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/PublicId" }
...
    VenueId:
      name: venueId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/PublicId" }
...
  responses:
    Unauthorized:
      description: Authentication is required.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    Forbidden:
      description: The active identity cannot access this workspace resource.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    Conflict:
      description: State or reservation conflict.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    PayloadTooLarge:
      description: JSON request body exceeds the global 1 MB limit.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
...
  schemas:
    B2Object:
      type: object
      additionalProperties: true
...
    TenantRole:
      type: object
      required: [id, name, templateCode, immutable, permissions]
      properties:
        id: { $ref: "#/components/schemas/PublicId" }
        name: { type: string }
        templateCode: { type: [string, "null"] }
        immutable: { type: boolean }
        permissions:
          type: array
          items: { type: string }
...
    PublicId:
      type: string
      pattern: "^[A-Za-z0-9_-]{22}$"
      description: Deterministically encrypted reference; internal numeric key is never exposed.
...
    ApiError:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string }
        message: { type: string }
        details: {}
        requestId: { type: string }
...
    Workspace:
      type: object
      required: [tenantId, name, slug, status, role, membershipId, assignedVenueIds]
      properties:
        tenantId: { $ref: "#/components/schemas/PublicId" }
        name: { type: string }
        slug: { type: string }
        status: { type: string }
        role: { type: string, enum: [PRIMARY_OWNER, OWNER, STAFF] }
        membershipId: { $ref: "#/components/schemas/PublicId" }
        assignedVenueIds:
          type: array
          items: { $ref: "#/components/schemas/PublicId" }
    WorkspaceMember:
      type: object
      required:
        [
          membershipId,
          userId,
          name,
          email,
          role,
          status,
          assignedVenueIds,
          tenantRoleId,
          tenantRoleName,
          permissions,
        ]
      properties:
        membershipId: { $ref: "#/components/schemas/PublicId" }
        userId: { $ref: "#/components/schemas/PublicId" }
        name: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [PRIMARY_OWNER, OWNER, STAFF] }
        status: { type: string }
        tenantRoleId:
          oneOf:
            - $ref: "#/components/schemas/PublicId"
            - type: "null"
        tenantRoleName: { type: [string, "null"] }
        permissions:
          type: array
          items: { type: string }
        assignedVenueIds:
          type: array
          items: { $ref: "#/components/schemas/PublicId" }
...
    CreateSignedUploadInput:
      type: object
      required: [tenantId, fileName, mimeType, byteSize]
      properties:
        tenantId: { $ref: "#/components/schemas/PublicId" }
        fileName: { type: string, minLength: 1, maxLength: 255 }
        mimeType:
          type: string
          enum: [image/webp, image/jpeg, image/png]
        byteSize: { type: integer, minimum: 1, maximum: 10485760 }
    SignedUpload:
      type: object
      required: [storageKey, uploadUrl, expiresInSeconds]
      properties:
        storageKey: { type: string }
        uploadUrl: { type: string, format: uri }
        expiresInSeconds: { type: integer }
    CompleteVenueMediaInput:
      type: object
      required: [tenantId, storageKey, mimeType, byteSize, altText, purpose]
      properties:
        tenantId: { $ref: "#/components/schemas/PublicId" }
        storageKey: { type: string, maxLength: 255 }
        mimeType: { type: string, enum: [image/webp, image/jpeg, image/png] }
        byteSize: { type: integer, minimum: 1, maximum: 10485760 }
        altText: { type: string, minLength: 3, maxLength: 150 }
        purpose: { type: string, enum: [COVER, GALLERY] }
...